+++ internal/modules/parking_session/parking_session.dto.go
@@ -3,9 +3,9 @@ package parking_session
 import "time"
 
 type ParkingSessionResponse struct {
-	ID          uint       `json:"id"`
-	LotID       uint       `json:"lot_id"`
-	SlotID      *uint      `json:"slot_id,omitempty"`
+	ID          uint64     `json:"id"`
+	LotID       uint64     `json:"lot_id"`
+	SlotID      *uint64    `json:"slot_id,omitempty"`
 	CardUID     string     `json:"card_uid"`
 	CardType    string     `json:"card_type"`
 	PlateNumber string     `json:"plate_number"`
@@ -16,9 +16,9 @@ type ParkingSessionResponse struct {
 }
 
 type ManageParkingSessionResponse struct {
-	ID          uint       `json:"id"`
-	LotID       uint       `json:"lot_id"`
-	SlotID      *uint      `json:"slot_id,omitempty"`
+	ID          uint64     `json:"id"`
+	LotID       uint64     `json:"lot_id"`
+	SlotID      *uint64    `json:"slot_id,omitempty"`
 	CardUID     string     `json:"card_uid"`
 	CardType    string     `json:"card_type"`
 	PlateNumber string     `json:"plate_number"`
@@ -47,18 +47,18 @@ type ManageParkingSessionListResponse struct {
 }
 
 type CreateParkingSessionInput struct {
-	LotID       uint   `json:"lot_id"`
+	LotID       uint64 `json:"lot_id"`
 	CardUID     string `json:"card_uid"`
 	CardType    string `json:"card_type"`
 	PlateNumber string `json:"plate_number"`
 }
 
 type AssignSlotInput struct {
-	SessionID uint `json:"session_id"`
-	SlotID    uint `json:"slot_id"`
+	SessionID uint64 `json:"session_id"`
+	SlotID    uint64 `json:"slot_id"`
 }
 
 type FinishParkingSessionInput struct {
-	SessionID uint  `json:"session_id"`
-	Fee       int64 `json:"fee"`
+	SessionID uint64 `json:"session_id"`
+	Fee       int64  `json:"fee"`
 }
